requirements-dev.txt
appdirs==1.4.4
asgiref==3.2.10
astroid==2.4.2
certifi==2020.6.20
colorama==0.4.3
distlib==0.3.1
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.1.1
filelock==3.0.12
isort==5.5.2
lazy-object-proxy==1.4.3
mccabe==0.6.1
pipenv==2020.8.13
pylint==2.6.0
pypyodbc==1.3.5.2
python-decouple==3.3
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
static3==0.7.0
toml==0.10.1
virtualenv==20.0.31
virtualenv-clone==0.5.4
whitenoise==5.2.0
wrapt==1.12.1
django-bootstrap-form
django-cpf
django-localflavor
